requirements.txt
absl-py==1.0.0
astor==0.8.1
cached-property==1.5.2
cycler==0.11.0
Cython==0.29.28
dataclasses==0.8
decorator==4.4.2
gast==0.5.3
google-pasta==0.2.0
grpcio==1.45.0
h5py==3.1.0
imageio==2.16.0
importlib-metadata==4.8.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.6
matplotlib==3.2.1
networkx==2.5.1
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
pkg-resources==0.0.0
protobuf==3.19.4
pyparsing==3.0.7
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==6.0
scikit-image==0.17.2
scipy==1.4.1
six==1.16.0
tabulate==0.8.7
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tifffile==2020.9.3
tqdm==4.46.1
typing-extensions==4.1.1
Wand==0.6.7
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.6.0
